Revert "Feature/refactor authanduser"
