+++ app/create-account.tsx
@@ -1,11 +1,8 @@
-import { Alert, KeyboardAvoidingView, Platform, View } from "react-native";
+import { Alert, View } from "react-native";
 
 import { useRouter } from "expo-router";
 
-import {
-  SafeAreaView,
-  useSafeAreaInsets,
-} from "react-native-safe-area-context";
+import { SafeAreaView } from "react-native-safe-area-context";
 
 import { useSignUp } from "@/src/features/auth/hooks/useSignUp";
 import AuthTitleSection from "@/src/features/auth/screens/AuthTitleSection";
@@ -16,8 +13,7 @@ import Divider from "@/src/shared/components/layout/Divider";
 
 export default function CreateAccount() {
   const router = useRouter();
-  const insets = useSafeAreaInsets();
-  const { mutateAsync: signUp } = useSignUp({
+  const { mutate: signUp } = useSignUp({
     onSuccess: (data) => {
       if (!data.session) {
         Alert.alert(
@@ -26,7 +22,7 @@ export default function CreateAccount() {
           [{ text: "OK", onPress: () => router.replace("/sign-in") }],
         );
       } else {
-        router.replace("/(protected)/setup-profile");
+        router.replace("/setup-profile");
       }
     },
     onError: (error) => {
@@ -37,8 +33,8 @@ export default function CreateAccount() {
     },
   });
 
-  const handleCreateAccount = async (data: AuthFormSection) => {
-    await signUp({
+  const handleCreateAccount = (data: AuthFormSection) => {
+    signUp({
       email: data.email,
       password: data.password,
     });
@@ -46,46 +42,40 @@ export default function CreateAccount() {
 
   return (
     <SafeAreaView className="flex-1 bg-white">
-      <KeyboardAvoidingView
-        behavior={Platform.OS === "ios" ? "padding" : undefined}
-        keyboardVerticalOffset={insets.top}
-        className="flex-1"
-      >
-        <View className="mx-8 flex flex-1 flex-col items-start justify-center">
-          {/* Create Account Title */}
-          <AuthTitleSection
-            title="Create Account"
-            subtitle="Already have an account?"
-            link="/sign-in"
-            linkText="SignIn"
-          />
+      <View className="mx-8 flex flex-1 flex-col items-start justify-center">
+        {/* Create Account Title */}
+        <AuthTitleSection
+          title="Create Account"
+          subtitle="Already have an account?"
+          link="/sign-in"
+          linkText="SignIn"
+        />
 
-          {/* Create Account Form */}
-          <FormSection
-            emailConfig={{
-              name: "email",
-              placeholder: "Email",
-              textContentType: "emailAddress",
-              autoCapitalize: "none",
-            }}
-            passwordConfig={{
-              name: "password",
-              placeholder: "Password",
-              textContentType: "newPassword",
-              autoCapitalize: "none",
-              secureTextEntry: true,
-            }}
-            buttonTitle="Create Account"
-            onSubmit={handleCreateAccount}
-          />
+        {/* Create Account Form */}
+        <FormSection
+          emailConfig={{
+            name: "email",
+            placeholder: "Email",
+            textContentType: "emailAddress",
+            autoCapitalize: "none",
+          }}
+          passwordConfig={{
+            name: "password",
+            placeholder: "Password",
+            textContentType: "newPassword",
+            autoCapitalize: "none",
+            secureTextEntry: true,
+          }}
+          buttonTitle="Create Account"
+          onSubmit={handleCreateAccount}
+        />
 
-          {/* Divider */}
-          <Divider text="or" />
+        {/* Divider */}
+        <Divider text="or" />
 
-          {/* SignIn with Social Media */}
-          <SocialOauthSection title="SignUp" />
-        </View>
-      </KeyboardAvoidingView>
+        {/* SignIn with Social Media */}
+        <SocialOauthSection />
+      </View>
     </SafeAreaView>
   );
 }
