+++ app/sign-in.tsx
@@ -1,11 +1,8 @@
-import { Alert, KeyboardAvoidingView, Platform, View } from "react-native";
+import { Alert, View } from "react-native";
 
 import { useRouter } from "expo-router";
 
-import {
-  SafeAreaView,
-  useSafeAreaInsets,
-} from "react-native-safe-area-context";
+import { SafeAreaView } from "react-native-safe-area-context";
 
 import { useSignIn } from "@/src/features/auth/hooks/useSignIn";
 import AuthTitleSection from "@/src/features/auth/screens/AuthTitleSection";
@@ -16,10 +13,9 @@ import Divider from "@/src/shared/components/layout/Divider";
 
 export default function SignIn() {
   const router = useRouter();
-  const insets = useSafeAreaInsets();
   const { mutateAsync: signIn } = useSignIn({
     onSuccess: () => {
-      router.replace("/(protected)/(tabs)");
+      router.replace("/(tabs)");
     },
     onError: (error) => {
       const friendlyMessage = error.message.includes("invalid credentials")
@@ -38,46 +34,40 @@ export default function SignIn() {
 
   return (
     <SafeAreaView className="flex-1 bg-white">
-      <KeyboardAvoidingView
-        behavior={Platform.OS === "ios" ? "padding" : undefined}
-        keyboardVerticalOffset={insets.top}
-        className="flex-1"
-      >
-        <View className="mx-8 flex flex-1 flex-col items-start justify-center">
-          {/* SignIn Title */}
-          <AuthTitleSection
-            title="Sign In"
-            subtitle="New here?"
-            link="/create-account"
-            linkText="Create an account"
-          />
+      <View className="mx-8 flex flex-1 flex-col items-start justify-center">
+        {/* SignIn Title */}
+        <AuthTitleSection
+          title="SignIn"
+          subtitle="New here?"
+          link="/create-account"
+          linkText="Create an account"
+        />
 
-          {/* SignIn Form */}
-          <FormSection
-            emailConfig={{
-              name: "email",
-              placeholder: "Email",
-              textContentType: "emailAddress",
-              autoCapitalize: "none",
-            }}
-            passwordConfig={{
-              name: "password",
-              placeholder: "Password",
-              textContentType: "password",
-              autoCapitalize: "none",
-              secureTextEntry: true,
-            }}
-            buttonTitle="Sign In"
-            onSubmit={handleSignIn}
-          />
+        {/* SignIn Form */}
+        <FormSection
+          emailConfig={{
+            name: "email",
+            placeholder: "Email",
+            textContentType: "emailAddress",
+            autoCapitalize: "none",
+          }}
+          passwordConfig={{
+            name: "password",
+            placeholder: "Password",
+            textContentType: "password",
+            autoCapitalize: "none",
+            secureTextEntry: true,
+          }}
+          buttonTitle="SignIn"
+          onSubmit={handleSignIn}
+        />
 
-          {/* Divider */}
-          <Divider text="or" />
+        {/* Divider */}
+        <Divider text="or" />
 
-          {/* SignIn with Social Media */}
-          <SocialOauthSection title="Sign In" />
-        </View>
-      </KeyboardAvoidingView>
+        {/* SignIn with Social Media */}
+        <SocialOauthSection />
+      </View>
     </SafeAreaView>
   );
 }
