+++ src/features/auth/screens/SocialOauthSection.tsx
@@ -4,12 +4,12 @@ import { Text } from "@react-navigation/elements";
 
 import SocialMediaButton from "../components/SocialMediaButton";
 
-export default function SocialOauthSection({ title }: { title: string }) {
+export default function SocialOauthSection() {
   return (
     <View className="flex w-full flex-col items-center justify-center gap-4">
       <View className="flex w-full flex-col items-center justify-center gap-2">
         <Text className="text-zinc-700">
-          {title} with your favorite social media account.
+          SignIn with your favorite social media account.
         </Text>
       </View>
       <View className="flex w-full flex-col items-center gap-4">
