+++ src/features/users/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./useCheckUserId";
 export * from "./useCreateProfile";
 export * from "./useProfile";
-export * from "./useProfileSetup";
