+++ src/features/users/index.ts
@@ -1,5 +1,3 @@
 export * from "./api";
 export * from "./hooks";
-export * from "./screens";
 export * from "./types";
-export * from "./utils";
