--- src/features/users/screens/index.ts
@@ -1,1 +1,0 @@
-export * from "./SetupProfileScreen";
