--- src/features/users/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./profile-setup.utils";
